chore: tmp state for moving from swagger to oas3

+++ src/main/java/com/kryptokrauts/aeternity/sdk/service/info/domain/KeyBlockResult.java
@@ -31,7 +31,7 @@ public class KeyBlockResult extends GenericResultObject<KeyBlock, KeyBlockResult
 
   private BigInteger target;
 
-  @Default private List<Integer> pow = new LinkedList<Integer>();
+  @Default private List<BigInteger> pow = new LinkedList<BigInteger>();
 
   private BigInteger nonce;
